requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.6.0
chromadb>=0.4.22
pdfplumber>=0.10.3
python-docx>=1.1.0
beautifulsoup4>=4.12.2
tiktoken>=0.5.2
pytest>=7.4.3
anyio>=4.0.0
httpx>=0.25.2
rank-bm25>=0.2.2
